Added the ability to add default rounds to the database

+++ app/src/main/java/eywa/projectcodex/database/entities/RoundArrowCount.kt
@@ -8,7 +8,10 @@ import androidx.room.Entity
 @Entity(tableName = "round_arrow_counts", primaryKeys = ["roundId", "distanceNumber"])
 data class RoundArrowCount(
         val roundId: Int,
+        /**
+         * distanceNumber 1 is the first distance shot
+         */
         val distanceNumber: Int,
-        val faceSizeInCm: Int,
+        val faceSizeInCm: Double,
         val arrowCount: Int
 )
